Fix for ArrayDelete
ArrayDelete crashes when called on certain 2D arrays. This occurs when the length passed into the instruction is longer than the dimension along which you are deleting. For example, deleting 3 "elements" (rows in this case) from a 3 x 2 array starting at row index 1. Currently, ArrayDelete will attempt to create an output array of size 3 x -1, which causes a crash when we call arrayOut->ResizeDimensions.

The fix is to account for the lengths of each dimension when computing endIndex and deletedPortionLength.

I also moved arrayOutLength into the if  (rank == 1) branch since it is only used there and caused confusion when reading the code.